`sct_register_to_template`: Move `register` functions into `core.py`
